New dict literals now support headers with the same syntax as blocks
