+++ src/aya/parser/tokens/DictToken.java
@@ -1,6 +1,7 @@
 package aya.parser.tokens;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 
 import aya.exceptions.parser.ParserException;
 import aya.exceptions.parser.SyntaxError;
@@ -8,10 +9,16 @@ import aya.instruction.DictLiteralInstruction;
 import aya.instruction.EmptyDictLiteralInstruction;
 import aya.instruction.Instruction;
 import aya.instruction.InstructionStack;
+import aya.instruction.variable.assignment.Assignment;
 import aya.obj.block.BlockUtils;
+import aya.obj.block.StaticBlock;
+import aya.obj.dict.Dict;
+import aya.obj.symbol.Symbol;
+import aya.parser.HeaderUtils;
 import aya.parser.Parser;
 import aya.parser.SourceStringRef;
 import aya.parser.token.TokenQueue;
+import aya.util.Triple;
 
 public class DictToken extends CollectionToken {
 
@@ -24,43 +31,32 @@ public class DictToken extends CollectionToken {
 		//Split Tokens where there are commas
 		ArrayList<TokenQueue> blockData = splitCommas(col);
 		
+		// No header, normal dict literal
 		if (blockData.size() == 1) {
 			InstructionStack instructions = Parser.generate(blockData.get(0));
+
+			// Optimization for empty dict literal
 			if (instructions.size() == 0) {
 				return EmptyDictLiteralInstruction.INSTANCE;
 			} else {
 				return new DictLiteralInstruction(this.getSourceStringRef(), BlockUtils.fromIS(instructions));
 			}
+
 		} else if (blockData.size() == 2) {
 			TokenQueue header = blockData.get(0);
-
-			// Single number in header, create a dict factory with a capture
-			if (header.size() == 1 && header.peek() instanceof NumberToken) {
-
-				NumberToken nt = (NumberToken)header.peek();
-				int n = 0;
-				try {
-					n = nt.numValue().toInt();
-				} catch (NumberFormatException e) {
-					throw new SyntaxError(nt + " is not a valid number in the blockEvaluator header", nt.getSourceStringRef());
-				}
-
-				if (n < 1) {
-					throw new SyntaxError("Cannot capture less than 1 elements from outer stack in a dict literal", nt.getSourceStringRef());
-				}
-				InstructionStack instructions = Parser.generate(blockData.get(1));
-				if (n == 0 && instructions.isEmpty()) {
-					return EmptyDictLiteralInstruction.INSTANCE;
-				} else {
-					return new DictLiteralInstruction(this.getSourceStringRef(), BlockUtils.fromIS(instructions), n);
-				}
+			
+			if (header.size() > 0) {
+				InstructionStack main_instructions = Parser.generate(blockData.get(1));
+				Triple<ArrayList<Assignment>, Dict, HashMap<Symbol, StaticBlock>> p = HeaderUtils.generateBlockHeader(blockData.get(0));
+				StaticBlock blk = BlockUtils.fromIS(main_instructions, p.second(), p.first());
+				return new DictLiteralInstruction(this.getSourceStringRef(), blk, p.third());
 			} else {
-				throw new SyntaxError("dict headers not supported (yet)", source);
+				throw new SyntaxError("Empty header not allowed in dict literal", source);
 			}
-
 		} else {
-			throw new SyntaxError("Dict headers not supported (yet)", source);
+			throw new SyntaxError("Dict literal has too many parts", source);
 		}
+
 	}
 
 	@Override
